ai/requirements.txt
fastapi
uvicorn[standard]
crewai==1.13.0
google-genai
chromadb==1.1.0
playwright==1.43.0
beautifulsoup4
feedparser
newsapi-python
fpdf2
python-telegram-bot==21.1.1
apscheduler
requests
python-dotenv
lxml
pandas
plotly
